# Add consumption state to the traffic-manager