--- a/cmd/traffic/cmd/manager/state/session.go
+++ b/cmd/traffic/cmd/manager/state/session.go
@@ -15,14 +15,18 @@ import (
 	"github.com/telepresenceio/telepresence/v2/pkg/tunnel"
 )
 
+const agentSessionIDPrefix = "agent:"
+
 type SessionState interface {
 	Cancel()
+	AwaitingBidiMapOwnerSessionID(stream tunnel.Stream) string
+	ConsumptionMetrics() *SessionConsumptionMetrics
 	Done() <-chan struct{}
 	LastMarked() time.Time
 	SetLastMarked(lastMarked time.Time)
 	Dials() <-chan *rpc.DialRequest
 	EstablishBidiPipe(context.Context, tunnel.Stream) (tunnel.Endpoint, error)
-	OnConnect(context.Context, tunnel.Stream, *int32) (tunnel.Endpoint, error)
+	OnConnect(context.Context, tunnel.Stream, *int32, *SessionConsumptionMetrics) (tunnel.Endpoint, error)
 }
 
 type awaitingBidiPipe struct {
@@ -38,6 +42,11 @@ type sessionState struct {
 	lastMarked          time.Time
 	awaitingBidiPipeMap map[tunnel.ConnID]*awaitingBidiPipe
 	dials               chan *rpc.DialRequest
+	consumptionMetrics  *SessionConsumptionMetrics
+}
+
+func (ss *sessionState) ConsumptionMetrics() *SessionConsumptionMetrics {
+	return ss.consumptionMetrics
 }
 
 // EstablishBidiPipe registers the given stream as waiting for a matching stream to arrive in a call
@@ -86,12 +95,27 @@ func (ss *sessionState) EstablishBidiPipe(ctx context.Context, stream tunnel.Str
 	}
 }
 
+func (ss *sessionState) AwaitingBidiMapOwnerSessionID(stream tunnel.Stream) string {
+	ss.Lock()
+	defer ss.Unlock()
+	if abp, ok := ss.awaitingBidiPipeMap[stream.ID()]; ok {
+		return abp.stream.SessionID()
+	}
+	return ""
+}
+
 // OnConnect checks if a stream is waiting for the given stream to arrive in order to create a BidiPipe.
 // If that's the case, the BidiPipe is created, started, and returned by both this method and the EstablishBidiPipe
 // method that registered the waiting stream. Otherwise, this method returns nil.
-func (ss *sessionState) OnConnect(_ context.Context, stream tunnel.Stream, counter *int32) (tunnel.Endpoint, error) {
+func (ss *sessionState) OnConnect(
+	ctx context.Context,
+	stream tunnel.Stream,
+	counter *int32,
+	consumptionMetrics *SessionConsumptionMetrics,
+) (tunnel.Endpoint, error) {
 	id := stream.ID()
 	ss.Lock()
+	// abp is a session corresponding to an end user machine
 	abp, ok := ss.awaitingBidiPipeMap[id]
 	if ok {
 		delete(ss.awaitingBidiPipeMap, id)
@@ -102,7 +126,13 @@ func (ss *sessionState) OnConnect(_ context.Context, stream tunnel.Stream, count
 		return nil, nil
 	}
 	name := fmt.Sprintf("%s: session %s -> %s", id, abp.stream.SessionID(), stream.SessionID())
-	bidiPipe := tunnel.NewBidiPipe(abp.stream, stream, name, counter)
+	tunnelProbes := &tunnel.BidiPipeProbes{}
+	if consumptionMetrics != nil {
+		tunnelProbes.BytesProbeA = consumptionMetrics.FromClientBytesChan
+		tunnelProbes.BytesProbeB = consumptionMetrics.ToClientBytesChan
+	}
+
+	bidiPipe := tunnel.NewBidiPipe(abp.stream, stream, name, counter, tunnelProbes)
 	bidiPipe.Start(abp.ctx)
 
 	defer close(abp.bidiPipeCh)
@@ -138,10 +168,11 @@ func (ss *sessionState) SetLastMarked(lastMarked time.Time) {
 func newSessionState(ctx context.Context, now time.Time) sessionState {
 	ctx, cancel := context.WithCancel(ctx)
 	return sessionState{
-		doneCh:     ctx.Done(),
-		cancel:     cancel,
-		lastMarked: now,
-		dials:      make(chan *rpc.DialRequest),
+		doneCh:             ctx.Done(),
+		cancel:             cancel,
+		lastMarked:         now,
+		dials:              make(chan *rpc.DialRequest),
+		consumptionMetrics: NewSessionConsumptionMetrics(),
 	}
 }