## Description

This PR extends the traffic manager state to include consumption metrics, so we can later expose them to prometheus.

Benchmark:

**Before**
```cli
➜  ~ wrk -t12 -c400 -d30s http://echo-easy.blue
Running 30s test @ http://echo-easy.blue
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    62.96ms   12.13ms 276.85ms   88.41%
    Req/Sec   319.44    134.07   666.00     66.77%
  114157 requests in 30.09s, 20.03MB read
  Socket errors: connect 157, read 0, write 0, timeout 0
Requests/sec:   3794.00
Transfer/sec:    681.73KB
```
**After**
```cli
~ wrk -t12 -c400 -d30s http://echo-easy.blue
Running 30s test @ http://echo-easy.blue
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    64.58ms   15.99ms 322.08ms   92.70%
    Req/Sec   311.97    185.42   670.00     58.31%
  111306 requests in 30.06s, 19.53MB read
  Socket errors: connect 157, read 0, write 0, timeout 0
Requests/sec:   3702.29
Transfer/sec:    665.26KB
```

No major difference.
## Checklist

<!--
  Please review the requirements for each checkbox, and check them
  off (change "[ ]" to "[x]") as you verify that they are complete.
-->

 - [ ] I made sure to add any docs changes required for my change (including release notes).
 - [x] My change is adequately tested.
 - [ ] I updated `DEVELOPING.md` with any special dev tricks I had to use to work on this code efficiently.
 - [ ] I updated `TELEMETRY.md` if I added, changed, or removed a metric name.
 - [ ] Once my PR is ready to have integration tests ran, I posted the PR in #telepresence-dev in the datawire-oss slack so that the "ok to test" label can be applied.