@josecv 2023-07-28
Also I think from a design perspective I'd feel much better if the channels were private and you just had methods to report one packet.